Integrated User Input Test Into Final User Input
