+++ client/src/components/UserInput/Step2Detailed.js
