+++ client/src/components/UserInput/Step7Detailed.js
@@ -1,27 +1,26 @@
 import React from 'react';
 
-export default class Step4Detailed extends React.Component {
-  next = e => {
-    e.preventDefault();
-    this.props.setOrders();
-    this.props.setOrderCost();
-    this.props.nextStep();
-  };
-
+export default class Step7Detailed extends React.Component {
   back = e => {
     e.preventDefault();
+    this.props.resetNameCost();
     this.props.prevStep();
   };
 
+  save = e => {
+    e.preventDefault();
+    alert("Saved");
+  }
+
   show = input => e =>{
     e.preventDefault();
-    input.orders.map((list) => (
+    input.names.map((list) => (
       console.log(list)
       ))
   }
 
   render(){ 
-    const { DetailedSplit, changeOrders, changeOrderCost } = this.props;
+    const { DetailedSplit } = this.props;
     return(
       <div className="row">
         <div className="col s12 m12 l12">
@@ -34,48 +33,38 @@ export default class Step4Detailed extends React.Component {
               </button>
             </div>
             <div className="card-content white-text">
-            <table className="highlight centered">
+              <table className="highlight centered">
                 <thead>
                   <tr>
-                    <th>Order Number</th>
-                    <th>Order Names</th>
-                    <th>Order Cost</th>
+                    <th>Names</th>
+                    <th>Payment</th>
                   </tr>
                 </thead>
                 <tbody>
-                  {DetailedSplit.orders.map((list, index) => (
+                  {DetailedSplit.names.map((list, index) => (
                     <tr key = {index}>
                       <td>
-                        {list.number}
+                        {list.name}
                       </td>
                       <td>
-                        <form>
-                          <input type="text" placeholder="Insert Order"
-                            defaultValue={list.order} onChange={changeOrders(index)}/>
-                        </form>
-                      </td> 
-                      <td>
-                        <form>
-                          <input type="number" min="0" step="0.01" placeholder="Insert Cost" className="validate"
-                            defaultValue={list.cost} onChange={changeOrderCost(index)}/>
-                        </form>
-                      </td> 
+                        ${list.cost}
+                      </td>
                     </tr>
                   ))}
                 </tbody>
               </table>
               <br/>
               <button className="btn waves-effect waves-light float-right"
-                type="submit" name="action" onClick={this.next}>
-                Next
-                <i className="material-icons right">navigate_next</i>
+                type="submit" name="action" onClick = {this.save}>
+                Save
+                <i className="material-icons right">save</i>
               </button>
-              <hr/>
+              {/* <hr/>
               <button className="btn waves-effect waves-light float-right"
                 type="submit" name="action" onClick={this.show(DetailedSplit)}>
                 Show
                 <i className="material-icons right">navigate_next</i>
-              </button>
+              </button> */}
             </div>
           </div>
         </div>
